1581021: Decode error from unicode passwords
Updated username/password read from config to unicode

Related 158022: username/password sent through BasicAuth as bytes
