+++ tests/test_rhevm.py
@@ -1,3 +1,5 @@
+# -*- coding: utf-8 -*-
+
 from __future__ import print_function
 """
 Test of RHEV-M virtualization backend.
@@ -104,7 +106,7 @@ class TestRhevM(TestBase):
 
     def setUp(self):
         config = self.create_config(name='test', wrapper=None, type='rhevm', server='localhost', username='username',
-                        password='password', owner='owner', env='env')
+                        password=u'1€345678', owner='owner', env='env')
         self.rhevm = Virt.from_config(self.logger, config, Datastore())
         self.rhevm.major_version = '3'
         self.rhevm.build_urls()
@@ -133,8 +135,8 @@ class TestRhevM(TestBase):
             call('https://localhost:8443/api/vms', auth=ANY, verify=ANY),
             call().raise_for_status(),
         ])
-        self.assertEqual(get.call_args[1]['auth'].username, 'username')
-        self.assertEqual(get.call_args[1]['auth'].password, 'password')
+        self.assertEqual(get.call_args[1]['auth'].username, u'username'.encode('utf-8'))
+        self.assertEqual(get.call_args[1]['auth'].password, u'1€345678'.encode('utf-8'))
 
     @patch('requests.get')
     def test_connection_refused(self, get):
